requirements.txt
streamlit
google-generativeai
requests
python-docx
python-dotenv